build(rulesets): abandon props mangling
