+++ .eslintignore
@@ -1,6 +1,6 @@
 **/__fixtures__/**
 /test-harness/tests/
 /packages/*/dist
-/packages/rulesets/src/oas/schemas/compiled.ts
+/packages/rulesets/src/oas/schemas/validators.ts
 /packages/*/CHANGELOG.md
 packages/formatters/src/html/templates.ts
